+++ R/dimensions.R
@@ -125,6 +125,11 @@ st_set_dimensions = function(.x, which, values = NULL, point = NULL, names = NUL
 		d[[which]] = create_dimension(values = values, point = point %||% d[[which]]$point, ...)
 		if (! is.null(names) && length(names) == 1)
 			base::names(d)[which] = names
+		if (!is.matrix(values) && isTRUE(attr(d, "raster")$curvilinear)) {
+			# FIXME: there's much more that should be checked for curvilinear grids...
+			# https://github.com/r-spatial/stars/issues/460
+			attr(d, "raster")$curvilinear = FALSE 
+		}
 #		else if (inherits(values, "sfc"))
 #			base::names(d)[which] = "sfc"
 	} else if (!is.null(point) && is.logical(point)) {