+++ NEWS.md
@@ -1,6 +1,8 @@
 # version 0.5-4
 
-* argument `downsample` in `plot.stars()` and `st_as_stars.stars_proxy()` has the same effect (removed a one-offset between them).
+* export `st_downsample()`, e.g. to be used by `tmap`; https://github.com/r-tmap/tmap/issues/597
+
+* argument `downsample` in `plot.stars()` and `st_as_stars.stars_proxy()` and `st_downsample()` has the same effect (removed a one-offset between them).
 
 * `st_redimension()` works for curvilinear grids; #441